+++ src/Cronofy/Requests/CreateCalendarRequest.cs
@@ -20,9 +20,18 @@
         /// Gets or sets the name for the calendar.
         /// </summary>
         /// <value>
-        /// The name for the channel.
+        /// The name for the calendar.
         /// </value>
         [JsonProperty("name")]
         public string Name { get; set; }
+
+        /// <summary>
+        /// Gets or sets the color for the calendar.
+        /// </summary>
+        /// <value>
+        /// The color for the calendar.
+        /// </value>
+        [JsonProperty("color")]
+        public string Color { get; set; }
     }
 }