+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [0.14.0]
+
+ * Support for revoking profile authorization [#41]
+
+## [0.13.0]
+
+ * Support for Real-Time Scheduling and Add To Calendar hour format option [#40]
+
+## [0.12.0]
+
+ * Support for color with calendar creation and event upsert [#38]
+ * Helper to verify push notification HMACs [#37]
+
 ## [0.11.0]
 
  * Support for real time booking [#36]
@@ -81,6 +94,9 @@
 [0.9.0]: https://github.com/cronofy/cronofy-csharp/releases/tag/rel-0.9.0
 [0.10.0]: https://github.com/cronofy/cronofy-csharp/releases/tag/rel-0.10.0
 [0.11.0]: https://github.com/cronofy/cronofy-csharp/releases/tag/rel-0.11.0
+[0.12.0]: https://github.com/cronofy/cronofy-csharp/releases/tag/rel-0.12.0
+[0.13.0]: https://github.com/cronofy/cronofy-csharp/releases/tag/rel-0.13.0
+[0.14.0]: https://github.com/cronofy/cronofy-csharp/releases/tag/rel-0.14.0
 
 [#3]: https://github.com/cronofy/cronofy-csharp/pull/3
 [#10]: https://github.com/cronofy/cronofy-csharp/pull/10
@@ -100,3 +116,7 @@
 [#31]: https://github.com/cronofy/cronofy-csharp/pull/31
 [#34]: https://github.com/cronofy/cronofy-csharp/pull/34
 [#36]: https://github.com/cronofy/cronofy-csharp/pull/36
+[#37]: https://github.com/cronofy/cronofy-csharp/pull/37
+[#38]: https://github.com/cronofy/cronofy-csharp/pull/38
+[#40]: https://github.com/cronofy/cronofy-csharp/pull/40
+[#41]: https://github.com/cronofy/cronofy-csharp/pull/41